+++ src/CrudServiceProvider.php
@@ -4,6 +4,7 @@ namespace Backpack\CRUD;
 
 use Route;
 use Illuminate\Support\ServiceProvider;
+use Prologue\Alerts\Facades\Alert;
 
 class CrudServiceProvider extends ServiceProvider
 {
@@ -56,7 +57,11 @@ class CrudServiceProvider extends ServiceProvider
                     '--tag' => 'public',
                 ]);
             }
+        } else {
+            $this->checkLicenseCodeExists();
         }
+
+        $this->checkLicenseCodeExists();
 
         // use the vendor configuration file as fallback
         $this->mergeConfigFrom(
@@ -111,5 +116,18 @@ class CrudServiceProvider extends ServiceProvider
         }
 
         return false;
+    }
+
+    /**
+     * Check to to see if a license code exists.
+     * If it does not, throw a notification bubble.
+     *
+     * @return void
+     */
+    private function checkLicenseCodeExists()
+    {
+        if (!env('BACKPACK_LICENSE')) {
+            Alert::add("warning", "<strong>You're using unlicensed software.</strong> Please <a target='_blank' href='http://backpackforlaravel.com'>purchase a license code</a> to hide this message.");
+        }
     }
 }